+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+##
+- Limit number of preamble bits of RMT encoder to 30
+- Rename dcc_encoder_config_t::cutoutbit_duration to bidibit_duration
+
 ## 0.32.0
 - Individual timings for namespace dcc::rx and dcc::tx
 - Move time2bit to namespace dcc::rx